+++ Models/PMF/SimplePlantModels/StrumTree.cs
@@ -12,7 +12,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Models.PMF.Scrum
+namespace Models.PMF.SimplePlantModels
 {
     /// <summary>
     /// Data structure that contains information for a specific crop type in Scrum
